typedef cleanup

+++ 17__header-and-defines/main.cpp
@@ -9,7 +9,7 @@ using namespace std;
 
 int main()
 {
-  cout << "version number, major: " << Basic_VERSION_MAJOR << ", minor: " << (int)Basic_VERSION_MINOR << endl;
+  cout << "version number, major: " << Basic_VERSION_MAJOR << ", minor: " << Basic_VERSION_MINOR << endl;
 
   return 0;
 }
