+++ 30__extern-libraries/main.cpp
@@ -12,7 +12,7 @@ using namespace std;
 
 int main()
 {
-  cout << "version number, major: " << Basic_VERSION_MAJOR << ", minor: " << (int)Basic_VERSION_MINOR << endl;
+  cout << "version number, major: " << Basic_VERSION_MAJOR << ", minor: " << Basic_VERSION_MINOR << endl;
 
   auto a = new A();
